refactor: rename storage/gorm package to storage/postgres
