+++ internal/storage/postgres/postgres_test.go
@@ -1,4 +1,4 @@
-package gormstorage
+package postgres
 
 import (
 	"sync/atomic"
