+++ internal/storage/sqlite/sqlite.go
@@ -13,7 +13,7 @@ import (
 	"github.com/OCAP2/extension/v5/internal/database"
 	"github.com/OCAP2/extension/v5/internal/logging"
 	"github.com/OCAP2/extension/v5/pkg/core"
-	gormstorage "github.com/OCAP2/extension/v5/internal/storage/gorm"
+	pgstorage "github.com/OCAP2/extension/v5/internal/storage/postgres"
 
 	"gorm.io/gorm"
 )
@@ -26,7 +26,7 @@ type Config struct {
 
 // Backend wraps the GORM backend for SQLite-specific behavior.
 type Backend struct {
-	*gormstorage.Backend
+	*pgstorage.Backend
 	db       *gorm.DB
 	cfg      Config
 	log      *logging.SlogManager
@@ -40,7 +40,7 @@ func New(cfg Config, entityCache *cache.EntityCache, markerCache *cache.MarkerCa
 		return nil, fmt.Errorf("failed to create in-memory SQLite DB: %w", err)
 	}
 
-	gormBackend := gormstorage.New(gormstorage.Dependencies{
+	gormBackend := pgstorage.New(pgstorage.Dependencies{
 		DB:          db,
 		EntityCache: entityCache,
 		MarkerCache: markerCache,
